Refactoring organs to make card a link
Removed the view button from the card and made the card a link to the
organ detail page.
