+++ src/sass/main.scss
@@ -31,3 +31,4 @@
 
 @import 'pages/edit';
 @import 'pages/discover';
+@import 'pages/organ';
