+++ src/sass/pages/organ.scss
@@ -1,0 +1,7 @@
+.card-hover {
+    &:hover {
+        box-shadow:
+            0 10px 20px rgba(0, 0, 0, 0.12),
+            0 4px 8px rgba(0, 0, 0, 0.06);
+    }
+}
